Promote SqlErrorMapping and DbSpec to top-level classes
Both types appear in the protected hook contract of
AbstractCobolEsqlBackend (mapSqlException / buildJdbcUrl), and
SqlErrorMapping had already been made fully public so nesting no longer
restricted its visibility. Cursor stays nested: it is base-class-internal
bookkeeping and not exposed to the SPI.

DbSpec keeps its package-private constructor (instances are created only
by the base class's buildSpec) and its public getters.

Also rename SqlErrorMapping.ecpgCode to sqlCode: the field is written to
SQLCA's SQLCODE, and the role-based name avoids PostgreSQL-specific
wording in a contract that non-PostgreSQL backends implement. The
SqlCA.ECPG_* constants keep their names since the canonical values
themselves come from the ECPG-compatible vocabulary.
